+++ terraform/s3/main.tf
@@ -21,7 +21,7 @@ resource "aws_s3_bucket_cors_configuration" "upload-cors" {
   bucket = aws_s3_bucket.earthdatapub-upload.id
   cors_rule {
     allowed_headers = ["*"]
-    allowed_methods = ["PUT"]
+    allowed_methods = ["POST"]
     allowed_origins = ["*"]
     max_age_seconds = 3000
   }